Bug 1568831 - [devtools] Make eyedropper fission compatible. r=ladybenko.
The current implementation of the eyedropper was using drawWindow, which does
not handle remote frames.
To make the tool fission compatible, we re-use the screenshot-content actor which
do support remote frames. We had to do a few modification on it though so the
eyedropper would work the same as it does at the moment:
- add a `disableFlash` property to the screenshot options, so we don't play the
  flash animation for the eyedropper
- retrieve the page current zoom level from the screenshot-content actor so we
  can adjust the size of the resulting image
- split the `dpr` screenshot option into 2 distinct ones:
  - `snapshotScale`, which will be the scale passed to `drawSnapshot`. By default,
    this will be the dpr of the window, so the screenshot is as crisp as it can be.
  - `fileScale`, which will be the scale of the resulting file.
- we control the fileScale that is sent to the server with the `ignoreDprForFileScale`
  on the client. This is needed by the eyedropper so it doesn't draw images that are
  too big when the dpr is > 1.

With all those changes, browser_inspector_highlighter-eyedropper-frames.js is passing
with fission enabled, so we can remove the `fail-if` annotation it had.

Differential Revision: https://phabricator.services.mozilla.com/D103879

UltraBlame original commit: 0bfd8c9d71a62c46e1a1329f2db780e8b0c21ee3