Bug 1691349 - [devtools] Fix screenshot node on about:preferences. r=bomsy.
The context menu action was freezing because of an infinite
loop in getRect: since we weren't passing a boundary window,
the function uses the top window, which can be the windowRoot.
But then, we walk up through the window using window.parent,
which would never get us the windowRoot, hence the infinite loop.
We fix this by providing the boundary window to getReact.
A test is added to ensure we don't regress in the future.

Differential Revision: https://phabricator.services.mozilla.com/D104388

UltraBlame original commit: 7ed0dc85c25bacbb9dad3bb188bca00d6c41a520