Bug 1689297 - [devtools] Add cloned content process message resource. r=ochameau,devtools-backward-compat-reviewers.
This is a legacy-listener-only resource that is required for us to receive the
messages cloned from the content process in the browser console when the `devtools.browsertoolbox.fission`
pref is not enabled.
Since the messages are sent through the `consoleApiCall` event, we add a `clonedFromContentProcess`
flag so we can ignore legit console api messages, which are handled as another resource.

This let us remove the specific code in WebConnectionProxy.

Differential Revision: https://phabricator.services.mozilla.com/D103284

UltraBlame original commit: 241b91491e1104a67f6ea9b5f22fb55b59a4f3c9