Bug 1666487 - Move key frame interval state from VideoTrackEncoder to VP8TrackEncoder. r=bryce
This changes the type of the interval to be unit aware, and makes it a Maybe to
allow for denoting a value that has not yet been set. This also moves the state
to VP8TrackEncoder to reduce the amount of indirection.

Differential Revision: https://phabricator.services.mozilla.com/D95701

UltraBlame original commit: 6454bf741983a4c7fa74c22a23bed5c9b3717523