Bug 1666487 - Effectively disable encoding keyframes when on a tight time budget. r=bryce
Most importantly this disables the keyframe encoding, but it also reduces the
margin for when to trigger a skipped frame to 85%.

Suddenly encoding a keyframe causes a quality degradation since the codec is
trying to constrain itself to its configured bitrate. Skipping a frame is less
intrusive since it essentially reduces the output framerate but retains the
visible quality of the frames. This is especially noticable with large
resolutions where keyframes can become very blocky.

Differential Revision: https://phabricator.services.mozilla.com/D95707

UltraBlame original commit: 9c56a50cc6d3761f381b4c6615e44b697fdd89ba