Bug 1666487 - Don't encode vp8 with 0 bitrate. r=bryce
An application can configure videoBitsPerSecond to be <1000, leading to the
configuration of the vp8 encoder with rc_target_bitrate = 0. A target bitrate
of 0 makes the encoder skip all frames.

Without this we tripped an assert in a crashtest with the whole stack of patches
applied.

Differential Revision: https://phabricator.services.mozilla.com/D95708

UltraBlame original commit: dd2bab72c12dc0781030286ae629f9c153c51baa