Bug 1464268 - MozPromisify MediaRecorder::Session::Extract and let GatherBlob paths extract remaining data first. r=bryce
This ensures that DoSessionEndTask will stop the MediaEncoder first, or cancel
it if data is to be discarded. We also wait for all the data still buffered to
be encoded, muxed and extracted into the blob before proceeding with
"dataavailable".

Differential Revision: https://phabricator.services.mozilla.com/D95716

UltraBlame original commit: e43d3445f0d4a953272b19555473c5cc38f8895e