Bug 1678119 - Simplify SWGL image fast-paths since complex blending is no longer required. r=jrmuizel
Now that most of the complicated alpha-pass features such as clip-masking and anti-aliasing
are handled in SWGL during the blend stage, most of the fast-paths are identical and only call
swgl_commitTextureLinear in a tight loop. We can do a lot better here by just moving that loop
into SWGL, not only making it faster but removing all the redundant boiler-plate code out of
the shaders.

Differential Revision: https://phabricator.services.mozilla.com/D104536

UltraBlame original commit: 1602ec8392d600bc047953518542fa3fe3fafd83