Bug 1686652 - Clean up SIMD metadata and features. r=rhunt
This just tidies up the wormhole switch, introducing a getter for it
on the moduleEnv.  As a drive-by fix, the v128Enabled switch in the
code metadata is removed, since nobody's using it.  (At least I think
so.  It does not seem to be part of anything we use to determine
whether cached code is valid, for example.)

Differential Revision: https://phabricator.services.mozilla.com/D103693

UltraBlame original commit: 6c5025b25cde2534545e4d13d84800fe4b7ae29a