Bug 1691612 - Simplify the CompilationStencil preparation code. r=arai
Store the `max*DataLength` values on the StencilDelazificationSet so that it
can be used for multiple instantiations. Separate out the preparation of the
stencil from the gc-output structure to allow multiple instantiations. Move
the delazificationIndices creation to part of XDR decode process. Remove the
flag and simply re-run the `reserve` operations to avoid problems with
stencil, atom-cache, and gc-output potentially having different life-times.

Differential Revision: https://phabricator.services.mozilla.com/D104744

UltraBlame original commit: 4b1348088c2476304198a0b2a86050373ecf1217