Bug 1692476 - [devtools] Fix browser_console_webconsole_private_browsing.js permafail on beta. r=ochameau.
The test is failing on beta because there `devtools.browsertoolbox.fission` is
set to false, so we are getting content error messages from the process target,
and we excluded _all_ private messages for process targets, whereas we only
want to exclude cached private messages.

As I tried to modify the test to run it with and without devtools.browsertoolbox.fission,
this revealed another issue when the pref is true; since we directly get the messages
from the frame target, we don't exclude cached private messages.
We don't handle this here directly, but in Bug 1692838,  so we don't leave the
test failing on beta for too long.

Differential Revision: https://phabricator.services.mozilla.com/D105165

UltraBlame original commit: 1760b5f880413c9f4bf24522b6af6614fc547777