Bug 1690698 - [devtools] Remove performReload option and call reload from the frontend r=ochameau,bomsy,devtools-backward-compat-reviewers
Depends on D104424

performReload is not a regular configuration option as it is not persisted.
It is rather a flag for a single call to Target::reconfigure.
As we move configuration options outside of the target, it will be easier to drive the reload from the frontend.
The call sites using performReload are quite rare and only found in the netmonitor so it doesn't feel like this
requires a framework solution for now. We add explicit calls to TargetFront::reload() in spots where the netmonitor
used to pass performReload = true.

Differential Revision: https://phabricator.services.mozilla.com/D104891

UltraBlame original commit: b39b60e8120de117b8bfb08830fe4551002b2e5d