Bug 1691428 - Fix styling of unthemed menupopups and menuitems in Linux. r=dao
The background change matches Windows / Mac, and common sense too.

The customoptionstyling change is because that comment is just false,
although custom select is not enabled on Linux (though maybe it could
after this patch, will look into it separately).

Finally, give some padding to unthemed menuitems. GTK overrides this
padding with the theme one, but this allows having a sane default.

This is the padding I get with Adwaita, which is the default GTK theme.

Differential Revision: https://phabricator.services.mozilla.com/D104874

UltraBlame original commit: 1edcdfd085591e20290d3ffab70f9aa7c0031e33