Bug 1690340 - Part 5: Remove the menu separators from the developer tools menu. r=jdescottes
This removes the menu separators frmo the developer tools menu to conform to
the proton design.

We can safely remove the "devToolsEndSeparator" separator since it is no longer
used to append the "Work Offline" that is removed in Part 4.

We change the anchor element to be the "menu_devtools_remotedebugging" menu item
with the removal oof the "menu_devtools_separator" separator.

Differential Revision: https://phabricator.services.mozilla.com/D104904

UltraBlame original commit: 2260fd69931c4972bf6190ea80f340889da40972