Bug 1588801 - Add test coverage for JS files loaded from XPIs. r=mixedpuppy,robwu
JS files in packed webextensions (XPIs) are loaded via `nsJARChannel`,
which checks the content type of a file based on the extension by
calling `nsIMIMEService::GetTypeFromExtension()`.

We might not be able to guarantee the exact set of JS file extensions
because this method also asks the OS but it should still be reliable.
We allow `.js`, `.jsm` and `.mjs` (defined in `defaultMimeEntries` and
`extraMimeEntries`).

Differential Revision: https://phabricator.services.mozilla.com/D104450

UltraBlame original commit: e9fe3f02c06a8b1d07ebb08a17772d820b281a9b