Bug 1644018 Force to trigger popup crop to fit the screen when we have size change from Wayland;r=stransky
We need to trigger popupFrame->SetPopupPosition when we get size change from the Wayland popup
positioner - ie. window does not fit the screen. That will ensure that the
popup is cropped to fit the screen.

This patch also keeps the NativeMoveResizeWaylandPopupCallback connected to make sure
that the subsequent nsWindow::Show calls places the window correctly (when for
example the main window is moved to the screen edge - the menus can be flipped - and
we need to update the nsView for that to ensure the submenus are correctly placed.

We also use anchorRectAppUnits.ToNearestPixels to partially mirror behaviour of
nsView::CalcWidgetBounds.

Differential Revision: https://phabricator.services.mozilla.com/D105507

UltraBlame original commit: 29a6032df4108887a29d11301c6151fe2a9b956c