Bug 1676943, add telemetry for how often tab or subframe crash ui is presented to the user. In addition, add telemetry to indicate that the user is shown the unsubmitted crashes notification bar, r=mconley
The telemetry is triggered either when the user is shown the tab crashed page, or when the subframe notification is displayed and the crash occurs in an active tab, or when the user switches to a tab that has a subframe crash notification displayed. The telemetry only applies the first time the tab is switched to. The unsubmitted crashes notification bar appears after 10 minutes or so and is shown in the top-most browser window, if one exists, and telemetry applies once when this occurs.

Differential Revision: https://phabricator.services.mozilla.com/D100349

UltraBlame original commit: 465ee033cea59e97ecdd6d9cec5f1ec26f0563c2