Bug 1665942 - Collect userContextId from tab in _collectBaseTabData and stop collecting it in the session history collection; r=nika,geckoview-reviewers,agi
### Story

When a COOP+COEP page is opened, we change the remoteness of the tab. After
changing the remoteness, the tab is then restored by its TabState. However,
we haven't collected userContextId through the current mechanism.

To fix that, this patch changes the way to collect userContextId.

### Test Plan

Test: D98296

Depends on D98296

Differential Revision: https://phabricator.services.mozilla.com/D98468

UltraBlame original commit: 51013501829a03bdf25680e8c97c2fc72810a15d