Bug 1694229 - Update slow script warning visuals r=florian
This patch removes the wait button on the slow script warning, on the suspicion
that it is confusing to the user since it's redundant with the close button. It
also changes the text of the notification to blame the hanging tab.

Differential Revision: https://phabricator.services.mozilla.com/D106015

UltraBlame original commit: d082f53d882eefff9e1c7e5f33efdd175e238eaf