Bug 1694229 - Show a different notification if selected tab is hanging r=florian
As part of this, we need to detect that the currently showing notification does
not match the one what we want to display. This also fixes the case where we
show a notification for tab A, then switch to tab B which is also hanging, and
end up listing the title for tab A as hanging in the notification.

Differential Revision: https://phabricator.services.mozilla.com/D106305

UltraBlame original commit: cb3d9e8d32e637010bdb1ab6c652449e3f8d42ac