Bug 1678505 - Add an additional test case that touch events are handled in non root scroll container having overscroll-behavior: none. r=botond
overscroll-behavior-none-on-non-root.html is a test case that the document has
a 100vh scrollable element and its overscroll-behavior is `none`, thus all
touch events are consumed by the non-root scroll container. But with dynamic
toolbar, it will be notified as if it's consumed by the root container to move
the dynamic toolbar if the scroll position in the scrollable element is at the
bottom edge.

Differential Revision: https://phabricator.services.mozilla.com/D105712

UltraBlame original commit: 7cd5f2c1c3f491b22a1f04af1a1f3f530a49261a