Bug 1695714: Clear Android's debug app setting when cleaning up after a mach run --debug session for GeckoView; r=firefox-build-system-reviewers,andi,mhentges
We always set the app as Android's debug app, which causes Android to wait for
a debugger to connect to the app before allowing it to start. We should clear
that setting once the debugger exits so that other developer activities (such as
manual testing) aren't blocked on the need to connect a debugger.

Differential Revision: https://phabricator.services.mozilla.com/D106807

UltraBlame original commit: c315567f4bbfab0227fa19eb5e61ffa912e50642