Bug 1695861 - Add check for shared strings in PreWriteBarrierDuringFlattening r=sfink
We don't need to peform the barrier for shared (or permanent strings) because they are not collected. The normal barrier checks this already but this special flattening path doesn't.

The function had to move to Marking-inl.h because it depends on calling JSString::isPermanentAndMayBeShared which isn't always defined when Cell.h is included.

Differential Revision: https://phabricator.services.mozilla.com/D106887

UltraBlame original commit: a1cd8533287de96e27037dda9c5b44df4beb05af