Bug 1695185 - Use the WindowId of the owner document instead r=valentin,necko-reviewers,baku
The websocket uses the innerWindowId of the window where the script lives when
triggering events for devtools. This is may not be the same as the websocket
owner window as mentioned here https://searchfox.org/mozilla-central/rev/f1159268add2fd0959e9f91b474f5da74c90f305/dom/websocket/WebSocket.cpp#206-207,212
This seems to be the case when we have iframes within a page. The devtools
do not get the events for the websockets channels opened within iframes as
the innerWindowId is that of the top-level window.

`...->TopWindowContext()->InnerWindowId()` also always point to the top-level window.

This patch fixes these issues by switching to use the websocket owner window which
the iframe.

Note: The parent revsion contains the changes for the devtools fission work that
depends on this fix. i did this to make it easy to test and review.

Differential Revision: https://phabricator.services.mozilla.com/D106607

UltraBlame original commit: 6fbfd2ff0ee71e1d9383e58ac1470c25aed878ab