Bug 1696026 - Fix assertion failure during sweep marking r=sfink
This is asserting because we're trying to get the barrier tracer for a zone that doesn't have barriers enabled, but is in the GC marking state. The reason is that barriers have been disabled because for sweeping, while we are also marking in zones that haven't started sweeping yet. This is fine. I rearranged the code so it skips the assertion.

Differential Revision: https://phabricator.services.mozilla.com/D107213

UltraBlame original commit: 4e523057e0e5f071c1fc3624ef127d5535538113