Bug 1696471 - [devtools] Notify about content process target destruction via watcher's target-destroyed-form. r=nchevobbe,jdescottes
We are currently receiving tabDetached for content process targets
when we destroy them via:
  Services.ppmm.broadcastAsyncMessage("debug:destroy-process-script")
done from process-helper and we stop watching for process targets.
Tests now depends on this behavior.

So I'm trying to replicate this behavior in order to later get rid of tabDetached
in favor of target-destroyed-form.

Differential Revision: https://phabricator.services.mozilla.com/D107248

UltraBlame original commit: 1acce69206357362ce5ced84ee40ce580def00cc