Bug 1697311 - Use the right clip for subsequent mask layers. r=mstange
The code was correctly computing `currentLayerClipState` for the right
layer, but then it was calling PrepareImage with `clipState` rather than
`currentLayerClipState`, which in the case of masks only contains the
clip for the bottom layer. This can cause background-repeat to
over-extend, causing the issue.

Differential Revision: https://phabricator.services.mozilla.com/D107758

UltraBlame original commit: 350441a59119d44a6563e81e55ff55105d30d500