Bug 1697270 - Provide GeckoResult.allow() and .deny() r=owlish,droeh
GeckoResult uses the thread is instantiated on to run callbacks, so it's
important that the calling code builds GeckoResult on the same thread where
it's used to avoid suprising cross-thread effects.

The above assumption is broken when using ALLOW and DENY as these objects are
built on whatever thread happens to load the GeckoResult class in the JVM.

This change deprecates ALLOW and DENY and introdocues allow() and deny() which
build the GeckoResult instance on the same thread where they are called
avoiding any cross-thread problems.

Differential Revision: https://phabricator.services.mozilla.com/D107729

UltraBlame original commit: d487e57a3d702f22d9665695a1e716132c63986a