Bug 1697416. Fix helper_bug1299195.html when double tap to zoom is enabled on mac. r=botond
The test send two taps, which triggers a double tap zoom if it's enabled, and then the test doesn't get the expected events. So change the test to disable zooming and then enable the meta viewport pref so that we respect it on desktop.

Differential Revision: https://phabricator.services.mozilla.com/D107791

UltraBlame original commit: f3dea1cca2d5b083589cf3fc59b69696e1b70c19